// src/app/models/product/product.ts
export interface Product {
  id: number;
  name: string;
  description: string;
  price: string;
  quantity: string;
  imageUrl: string;

  // constructor(id: number, name: string, description: string, price: string, quantity: string, imageUrl: string) {
  //   this.id = id;
  //   this.name = name;
  //   this.description = description;
  //   this.price = price;
  //   this.quantity = quantity;
  //   this.imageUrl = imageUrl;
  // }
}